pyproject.toml
[tool.poetry]
name = "pneumonia_detect"
version = "0.1.0"
description = "Pneumonia Detection with Vision Transformers"
authors = ["Admin <admin@example.com>"]
readme = "README.md"
packages = [{include = "pneumonia_detect", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"

torch = "^2.2.0"
torchvision = "^0.17.0"

transformers = "^4.36.0"
datasets = "^2.14.0"
accelerate = ">=0.26.0"
evaluate = "^0.4.0"

numpy = "^1.24.0"
pandas = "^2.0.0"
scikit-learn = "^1.3.0"
imbalanced-learn = ">=0.12.0"
matplotlib = "^3.7.0"
pillow = "^10.0.0"

hydra-core = "^1.3.2"
dvc = "^3.0.0"
mlflow = "^2.5.0"
fire = "^0.7.1"
kaggle = "1.6.17"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pre-commit = "^3.3.3"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
ignore_missing_imports = true
